requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.2.2
discord.py==2.3.2
Django==3.2.23
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
onipkg_contrib @ git+https://github.com/Onimusic/onipkg_contrib.git@v1.2b
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
yarl==1.9.4
